Remove CSS files and update component structure
The App.css and index.css files have been deleted, leading to a major restructuring of the React components. A new "Template" component was added in the client/src/templates directory. The "App" component has been modified to incorporate this. In addition, changes were made to the prettier.xml file, removing the 'myRunOnSave' option.